+++ src/main/java/com/sparta/itsmine/domain/auction/dto/GetAuctionByProductResponseDto.java
@@ -1,6 +1,7 @@
 package com.sparta.itsmine.domain.auction.dto;
 
 import com.querydsl.core.annotations.QueryProjection;
+import com.sparta.itsmine.domain.product.utils.ProductStatus;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -11,10 +12,10 @@ public class GetAuctionByProductResponseDto {
 
     private Long userId;
     private Long productId;
-    private Long bidPrice;
+    private Integer bidPrice;
 
     @QueryProjection
-    public GetAuctionByProductResponseDto(Long productId, Long bidPrice, Long userId) {
+    public GetAuctionByProductResponseDto(Long productId, Integer bidPrice, Long userId) {
         this.productId = productId;
         this.bidPrice = bidPrice;
         this.userId = userId;
